Add subscribers to groups

+++ code/Server.Api/Controllers/QuestionSubscriptionController.cs
@@ -9,6 +9,8 @@ using Microsoft.AspNetCore.Mvc;
 using Server.Api.Dtos;
 using Server.Api.Models;
 using Server.Api.Repositories;
+using Microsoft.AspNetCore.SignalR;
+using ChatSample.Hubs;
 
 namespace Server.Api.Controllers
 {
@@ -18,12 +20,14 @@ namespace Server.Api.Controllers
     {
         private readonly IQuestionSubscriptionRepository _questionSubscriptionRepository;
         private readonly UserManager<User> _userManager;
+		private readonly IHubContext<ChatHub> _hubContext;
 
-        public QuestionSubscriptionController(IQuestionSubscriptionRepository repository, UserManager<User> userManager)
+        public QuestionSubscriptionController(IQuestionSubscriptionRepository repository, UserManager<User> userManager, IHubContext<ChatHub> hubContext)
         {
             _questionSubscriptionRepository = repository;
             _userManager = userManager;
-        }
+			_hubContext = hubContext;
+		}
         
         private async Task<IEnumerable<getQuestionSubscriptionDto>> _getQuestionSubscriptions()
         {
@@ -85,6 +89,7 @@ namespace Server.Api.Controllers
                     userId = user.Id,
                     questionId = dto.questionId,
                 };
+                await _hubContext.Groups.AddToGroupAsync(UserHandler.ConnectedIds[user.Id], "Question " + subscription.questionId.ToString());
                 await _questionSubscriptionRepository.createAsync(subscription);
                 return Ok(subscription);
             }
@@ -99,8 +104,19 @@ namespace Server.Api.Controllers
         {
             try
             {
-                await _questionSubscriptionRepository.deleteAsync(id);
-            }
+                QuestionSubscription sub = await _questionSubscriptionRepository.getAsync(id);
+                ClaimsPrincipal currentUser = HttpContext.User;
+				if (currentUser.HasClaim(c => c.Type == "username"))
+				{
+					string userEmail = currentUser.Claims.FirstOrDefault(c => c.Type == "username").Value;
+					User user = await _userManager.FindByEmailAsync(userEmail);
+
+					await _hubContext.Groups.RemoveFromGroupAsync(UserHandler.ConnectedIds[user.Id], "Question " + sub.questionId.ToString());
+					await _questionSubscriptionRepository.deleteAsync(id);
+				} else {
+					return Unauthorized();
+				}
+			}
             catch (System.Exception)
             {
                 return NotFound();
